fix test group names
