+++ onnxruntime/test/providers/cpu/activation/activation_op_test.cc
@@ -2,10 +2,38 @@
 // Licensed under the MIT License.
 
 #include "core/providers/cpu/activation/activations.h"
-#include "test/common/activation_test_utils.h"
+#include "gtest/gtest.h"
+#include "test/providers/provider_test_utils.h"
 
 namespace onnxruntime {
 namespace test {
+
+void TestUnaryElementwiseOp(const char* szOp, std::vector<float>& input_vals,
+                            std::function<float(float)> expected_func,
+                            const std::unordered_map<std::string, float> attribs = {},
+                            bool is_tensorrt_supported = true,
+                            int opset_version = 7) {
+  OpTester test(szOp, opset_version);
+
+  for (auto attr : attribs)
+    test.AddAttribute(attr.first, attr.second);
+
+  std::vector<int64_t> dims{(int64_t)input_vals.size()};
+
+  std::vector<float> expected_vals;
+  for (const auto& iv : input_vals)
+    expected_vals.push_back(expected_func(iv));
+
+  test.AddInput<float>("X", dims, input_vals);
+  test.AddOutput<float>("Y", dims, expected_vals);
+
+  // Disable TensorRT on unsupported tests
+  std::unordered_set<std::string> excluded_providers;
+  if (!is_tensorrt_supported) {
+    excluded_providers.insert(kTensorrtExecutionProvider);
+  }
+  test.Run(OpTester::ExpectResult::kExpectSuccess, "", excluded_providers);
+}
 
 std::vector<float> input_vals = {
     -1.0f, 0, 1.0f,                                              // normal input values for activation
@@ -161,7 +189,8 @@ TEST(ActivationOpTest, Softplus) {
                              return x + logf(expf(-x) + 1);
                            else
                              return logf(expf(x) + 1);
-                         }, {}, false);
+                         },
+                         {}, false);
 }
 
 TEST(ActivationOpTest, Softsign) {
